now with path compression
