+++ internal/bitset/bitset_test.go
@@ -509,87 +509,6 @@ func TestAsSlice(t *testing.T) {
 	}
 }
 
-func TestAsSliceFunc(t *testing.T) {
-	t.Parallel()
-	all := func(uint) bool { return true }
-	none := func(uint) bool { return false }
-	even := func(u uint) bool { return u%2 == 0 }
-	odd := func(u uint) bool { return u%2 == 1 }
-
-	testCases := []struct {
-		name string
-		//
-		set      []uint
-		buf      []uint
-		filter   func(uint) bool
-		wantData []uint
-	}{
-		{
-			name:     "null",
-			set:      []uint{},
-			buf:      make([]uint, 0, 512),
-			filter:   all,
-			wantData: []uint{},
-		},
-		{
-			name:     "zero",
-			set:      []uint{0},
-			buf:      make([]uint, 0, 512),
-			filter:   all,
-			wantData: []uint{0}, // bit #0 is set
-		},
-		{
-			name:     "1,5",
-			set:      []uint{1, 5},
-			buf:      make([]uint, 0, 512),
-			filter:   all,
-			wantData: []uint{1, 5},
-		},
-		{
-			name:     "many",
-			set:      []uint{1, 65, 130, 190, 250, 300, 380, 420, 480, 511},
-			buf:      make([]uint, 0, 512),
-			filter:   all,
-			wantData: []uint{1, 65, 130, 190, 250, 300, 380, 420, 480, 511},
-		},
-		{
-			name:     "none",
-			set:      []uint{1, 65, 130, 190, 250, 300, 380, 420, 480, 511},
-			buf:      make([]uint, 0, 512),
-			filter:   none,
-			wantData: []uint{},
-		},
-		{
-			name:     "even",
-			set:      []uint{1, 65, 130, 190, 250, 300, 380, 420, 480, 511},
-			buf:      make([]uint, 0, 512),
-			filter:   even,
-			wantData: []uint{130, 190, 250, 300, 380, 420, 480},
-		},
-		{
-			name:     "odd",
-			set:      []uint{1, 65, 130, 190, 250, 300, 380, 420, 480, 511},
-			buf:      make([]uint, 0, 512),
-			filter:   odd,
-			wantData: []uint{1, 65, 511},
-		},
-	}
-
-	for _, tc := range testCases {
-		var b BitSet
-		for _, u := range tc.set {
-			b = b.Set(u)
-		}
-
-		buf := b.AsSliceFunc(tc.buf, tc.filter)
-
-		if !slices.Equal(buf, tc.wantData) {
-			t.Errorf("AsSliceFunc, %s: returned buf is not equal as expected:\ngot:  %v\nwant: %v",
-				tc.name, buf, tc.wantData)
-		}
-	}
-}
-
 func TestCount(t *testing.T) {
 	t.Parallel()
 	var b BitSet
