+++ jsonify.go
@@ -1,7 +1,7 @@
+package bart
+
 // Copyright (c) 2024 Karl Gaissmaier
 // SPDX-License-Identifier: MIT
-
-package bart
 
 import (
 	"encoding/json"
@@ -64,7 +64,7 @@ func (n *node[V]) dumpListRec(parentIdx uint, path [16]byte, depth int, is4 bool
 	}
 
 	directKids := n.getKidsRec(parentIdx, path, depth, is4)
-	slices.SortFunc(directKids, cmpKidByPrefix[V])
+	slices.SortFunc(directKids, cmpKidByPrefix2[V])
 
 	nodes := make([]DumpListNode[V], 0, len(directKids))
 	for _, kid := range directKids {
