+++ jsonify_test.go
@@ -32,7 +32,7 @@ func TestJSONTableIsNil(t *testing.T) {
 		want: "null",
 	}
 
-	var tbl *Table2[any]
+	var tbl *Table[any]
 	checkJSON(t, tbl, tt)
 }
 
@@ -43,7 +43,7 @@ func TestJSONEmpty(t *testing.T) {
 		want: "{}",
 	}
 
-	tbl := new(Table2[any])
+	tbl := new(Table[any])
 	checkJSON(t, tbl, tt)
 }
 
@@ -57,7 +57,7 @@ func TestJSONDefaultRouteV4(t *testing.T) {
 		want: `{"ipv4":[{"cidr":"0.0.0.0/0","value":null}]}`,
 	}
 
-	tbl := new(Table2[any])
+	tbl := new(Table[any])
 	checkJSON(t, tbl, tt)
 }
 
@@ -71,7 +71,7 @@ func TestJSONDefaultRouteV6(t *testing.T) {
 		want: `{"ipv6":[{"cidr":"::/0","value":31337}]}`,
 	}
 
-	tbl := new(Table2[any])
+	tbl := new(Table[any])
 	checkJSON(t, tbl, tt)
 }
 
@@ -119,7 +119,7 @@ func TestJSONSampleV4(t *testing.T) {
 		want: `{"ipv4":[{"cidr":"10.0.0.0/8","value":null,"subnets":[{"cidr":"10.0.0.0/24","value":null},{"cidr":"10.0.1.0/24","value":null}]},{"cidr":"127.0.0.0/8","value":null,"subnets":[{"cidr":"127.0.0.1/32","value":null}]},{"cidr":"169.254.0.0/16","value":null},{"cidr":"172.16.0.0/12","value":null},{"cidr":"192.168.0.0/16","value":null,"subnets":[{"cidr":"192.168.1.0/24","value":null}]}]}`,
 	}
 
-	tbl := new(Table2[any])
+	tbl := new(Table[any])
 	checkJSON(t, tbl, tt)
 }
 
@@ -149,7 +149,7 @@ func TestJSONSampleV6(t *testing.T) {
 		want: `{"ipv6":[{"cidr":"::1/128","value":null},{"cidr":"2000::/3","value":null,"subnets":[{"cidr":"2001:db8::/32","value":null}]},{"cidr":"fe80::/10","value":null}]}`,
 	}
 
-	tbl := new(Table2[any])
+	tbl := new(Table[any])
 	checkJSON(t, tbl, tt)
 }
 
@@ -225,11 +225,11 @@ func TestJSONSample(t *testing.T) {
 		want: `{"ipv4":[{"cidr":"10.0.0.0/8","value":null,"subnets":[{"cidr":"10.0.0.0/24","value":null},{"cidr":"10.0.1.0/24","value":null}]},{"cidr":"127.0.0.0/8","value":31337,"subnets":[{"cidr":"127.0.0.1/32","value":"some string"}]},{"cidr":"169.254.0.0/16","value":3.14},{"cidr":"172.16.0.0/12","value":null},{"cidr":"192.168.0.0/16","value":["a","c","ff"],"subnets":[{"cidr":"192.168.1.0/24","value":"550e8400-e29b-41d4-a716-446655440000"}]}],"ipv6":[{"cidr":"::/0","value":null,"subnets":[{"cidr":"::1/128","value":null},{"cidr":"2000::/3","value":null,"subnets":[{"cidr":"2001:db8::/32","value":null}]},{"cidr":"fe80::/10","value":null}]}]}`,
 	}
 
-	tbl := new(Table2[any])
+	tbl := new(Table[any])
 	checkJSON(t, tbl, tt)
 }
 
-func checkJSON(t *testing.T, tbl *Table2[any], tt jsonTest) {
+func checkJSON(t *testing.T, tbl *Table[any], tt jsonTest) {
 	t.Helper()
 	for _, node := range tt.nodes {
 		tbl.Insert(node.cidr, node.value)
