+++ node.go
@@ -1,7 +1,7 @@
+package bart
+
 // Copyright (c) 2024 Karl Gaissmaier
 // SPDX-License-Identifier: MIT
-
-package bart
 
 import (
 	"net/netip"
@@ -32,110 +32,153 @@ var zeroPath [16]byte
 // The array slots are also not pre-allocated (alloted) as described
 // in the ART algorithm, but backtracking is used for the longest-prefix-match.
 //
-// The lookup is then slower, but this is the intended trade-off to prevent
-// memory consumption from exploding.
-//
-// The nodes can also be used in path compressd mode, this reduces the memory consumption
-// by almost an order of magnitude, but the updates (insert/delete) are slower,
-// but the search times remain comparable.
+// The sparse child array recursively spans the trie with a branching factor of 256
+// and also records path-compressed leaves in the free child slots.
 type node[V any] struct {
-	// prefixes contains the routes with payload V
+	// prefixes contains the routes as complete binary tree with payload V
 	prefixes sparse.Array[V]
 
 	// children, recursively spans the trie with a branching factor of 256
-	children sparse.Array[*node[V]]
-
-	// path compressed items
-	pathcomp sparse.Array[*pathItem[V]]
+	// the generic child with empty interface is a node (recursive) or
+	// a path compressed leaf (prefix and value).
+	children sparse.Array[interface{}]
 }
 
-// pathItem is prefix and value together
-type pathItem[V any] struct {
+// leaf is prefix and value together as path compressed child
+type leaf[V any] struct {
 	prefix netip.Prefix
 	value  V
 }
 
-// isEmpty returns true if node has neither prefixes nor children nor path compressed items.
-func (n *node[V]) isEmpty() bool {
-	return n.prefixes.Len() == 0 && n.children.Len() == 0 && n.pathcomp.Len() == 0
+// cloneValue, deep copy if v implements the Cloner interface.
+func cloneValue[V any](v V) V {
+	if k, ok := any(v).(Cloner[V]); ok {
+		return k.Clone()
+	}
+	return v
 }
 
-// numNodesRec, calculate the number of nodes under n.
-func (n *node[V]) numNodesRec() int {
+// cloneLeaf returns a copy of the leaf.
+// If the value implements the Cloner interface, the values are deeply copied.
+func (l *leaf[V]) cloneLeaf() *leaf[V] {
+	if l == nil {
+		return nil
+	}
+	return &leaf[V]{l.prefix, cloneValue(l.value)}
+}
+
+// isEmpty returns true if node has neither prefixes nor children
+func (n *node[V]) isEmpty() bool {
+	return n.prefixes.Len() == 0 && n.children.Len() == 0
+}
+
+// nodeAndLeafCount
+func (n *node[V]) nodeAndLeafCount() (nodes int, leaves int) {
+	for i := range n.children.AsSlice(make([]uint, 0, maxNodeChildren)) {
+		switch n.children.Items[i].(type) {
+		case *node[V]:
+			nodes++
+		case *leaf[V]:
+			leaves++
+		}
+	}
+	return
+}
+
+// nodeAndLeafCountRec, calculate the number of nodes and leaves under n, rec-descent.
+func (n *node[V]) nodeAndLeafCountRec() (int, int) {
 	if n == nil || n.isEmpty() {
-		return 0
+		return 0, 0
 	}
 
-	size := 1 // this node
+	nodes := 1 // this node
+	leaves := 0
+
 	for _, c := range n.children.Items {
-		size += c.numNodesRec()
+		switch k := c.(type) {
+		case *node[V]:
+			// rec-descent
+			ns, ls := k.nodeAndLeafCountRec()
+			nodes += ns
+			leaves += ls
+
+		case *leaf[V]:
+			leaves++
+		}
 	}
-	return size
+
+	return nodes, leaves
 }
 
 // insertAtDepth insert a prefix/val into a node tree at depth.
 // n must not be nil, prefix must be valid and already in canonical form.
 //
-// Required if a path compression has to be resolved because a new value
-// is added that collides with the compression. The prefix is then reinserted
-// further down at depth in the tree.
+// If a path compression has to be resolved because a new value is added
+// that collides with a leaf, the compressed leaf is then reinserted
+// one depth down in the node trie.
 func (n *node[V]) insertAtDepth(pfx netip.Prefix, val V, depth int) (exists bool) {
-	ip := pfx.Addr()
+	octets := pfx.Addr().AsSlice()
 	bits := pfx.Bits()
-	octets := ip.AsSlice()
 
-	lastOctetIdx := (bits - 1) / strideLen
-	lastOctet := octets[lastOctetIdx]
-	lastOctetBits := bits - (lastOctetIdx * strideLen)
+	// 10.0.0.0/8    -> 0
+	// 10.12.0.0/15  -> 1
+	// 10.12.0.0/16  -> 1
+	// 10.12.10.9/32 -> 3
+	sigOctetIdx := (bits - 1) / strideLen
 
-	// take the desired octets from prefix, starting at depth
-	for i := depth; i < lastOctetIdx; i++ {
-		addr := uint(octets[i])
+	// 10.0.0.0/8    -> 10
+	// 10.12.0.0/15  -> 12
+	// 10.12.0.0/16  -> 12
+	// 10.12.10.9/32 -> 9
+	sigOctet := octets[sigOctetIdx]
 
-		// descend down to next trie level
-		if n.children.Test(addr) {
-			n = n.children.MustGet(addr)
-			continue
-		}
+	// 10.0.0.0/8    -> 8
+	// 10.12.0.0/15  -> 7
+	// 10.12.0.0/16  -> 8
+	// 10.12.10.9/32 -> 8
+	sigOctetBits := bits - (sigOctetIdx * strideLen)
 
-		// no child found, look for path compressed item in slot
+	// find the proper trie node to insert prefix
+	// start with prefix octet at depth
+	for ; depth < sigOctetIdx; depth++ {
+		addr := uint(octets[depth])
 
-		if !n.pathcomp.Test(addr) {
+		if !n.children.Test(addr) {
 			// insert prefix path compressed
-			return n.pathcomp.InsertAt(addr, &pathItem[V]{pfx, val})
+			return n.children.InsertAt(addr, &leaf[V]{pfx, val})
 		}
 
-		// pathcomp slot is already occupied
-		pc := n.pathcomp.MustGet(addr)
+		// get the child: node or leaf
+		switch k := n.children.MustGet(addr).(type) {
+		case *node[V]:
+			// descend down to next trie level
+			n = k
+		case *leaf[V]:
+			// reached a path compressed prefix
+			// override value in slot if prefixes are equal
+			if k.prefix == pfx {
+				k.value = val
+				// exists
+				return true
+			}
 
-		// override prefix in slot if equal
-		if pc.prefix == pfx {
-			pc.value = val
-			return true
+			// create new node
+			// push the leaf down
+			// insert new child at cureent leaf position (addr)
+			// descend down, replace n with new child
+			c := new(node[V])
+			c.insertAtDepth(k.prefix, k.value, depth+1)
+
+			n.children.InsertAt(addr, c)
+			n = c
 		}
-
-		// free this pathcomp slot ...
-		// insert new intermdiate child ...
-		// shuffle down existing path-compressed prefix
-		// shuffle down prefix
-		n.pathcomp.DeleteAt(addr)
-
-		c := new(node[V])
-		n.children.InsertAt(addr, c)
-		n = c
-
-		// shuffle down
-		_ = n.insertAtDepth(pc.prefix, pc.value, depth+1)
-
-		// rec-descent monster ...
-		return n.insertAtDepth(pfx, val, depth+1)
 	}
 
-	// insert/override flattened prefix/val into node
-	return n.prefixes.InsertAt(pfxToIdx(lastOctet, lastOctetBits), val)
+	// last significant octet: insert/override prefix/val into node
+	return n.prefixes.InsertAt(pfxToIdx(sigOctet, sigOctetBits), val)
 }
 
-// purgeParents, dangling nodes after successful deletion
+// TODO, path compress after purging dangling leaves
 func (n *node[V]) purgeParents(parentStack []*node[V], childPath []byte) {
 	for i := len(parentStack) - 1; i >= 0; i-- {
 		if n.isEmpty() {
@@ -189,294 +232,6 @@ func (n *node[V]) lpmTest(idx uint) bool {
 	return false
 }
 
-// ### more complex functions than routing table lookups ###
-
-// eachLookupPrefix does an all prefix match in the 8-bit (stride) routing table
-// at this depth and calls yield() for any matching CIDR.
-func (n *node[V]) eachLookupPrefix(
-	path [16]byte,
-	depth int,
-	is4 bool,
-	octet byte,
-	pfxLen int,
-	yield func(netip.Prefix, V) bool,
-) bool {
-	// backtracking the CBT
-	for idx := pfxToIdx(octet, pfxLen); idx > 0; idx >>= 1 {
-		if n.prefixes.Test(idx) {
-			val := n.prefixes.MustGet(idx)
-			cidr, _ := cidrFromPath(path, depth, is4, idx)
-
-			if !yield(cidr, val) {
-				return false
-			}
-		}
-	}
-
-	return true
-}
-
-// eachSubnet calls yield() for any covered CIDR by parent prefix in natural CIDR sort order.
-func (n *node[V]) eachSubnet(
-	pfx netip.Prefix,
-	path [16]byte,
-	depth int,
-	is4 bool,
-	octet byte,
-	pfxLen int,
-	yield func(netip.Prefix, V) bool,
-) bool {
-	// ###############################################################
-	// 1. collect all indices in n covered by prefix
-	// ###############################################################
-	pfxFirstAddr := uint(octet)
-	pfxLastAddr := uint(octet | ^netMask(pfxLen))
-
-	allCoveredIndices := make([]uint, 0, maxNodePrefixes)
-
-	var idx uint
-	var ok bool
-	for {
-		if idx, ok = n.prefixes.NextSet(idx); !ok {
-			break
-		}
-
-		// idx is covered by prefix?
-		thisOctet, thisPfxLen := idxToPfx(idx)
-
-		thisFirstAddr := uint(thisOctet)
-		thisLastAddr := uint(thisOctet | ^netMask(thisPfxLen))
-
-		if thisFirstAddr >= pfxFirstAddr && thisLastAddr <= pfxLastAddr {
-			allCoveredIndices = append(allCoveredIndices, idx)
-		}
-
-		idx++
-	}
-
-	// sort indices in CIDR sort order
-	slices.SortFunc(allCoveredIndices, cmpIndexRank)
-
-	// ################################################################
-	// 2. collect all path compressed prefixes in n covered by prefix
-	// ################################################################
-
-	allCoveredPathCompSet := make([]bool, maxNodeChildren)
-
-	var addr uint
-	for {
-		if addr, ok = n.pathcomp.NextSet(addr); !ok {
-			break
-		}
-
-		// pathcomp addrs are sorted in indexRank order
-		if addr > pfxLastAddr {
-			break
-		}
-
-		if addr >= pfxFirstAddr {
-			pc := n.pathcomp.MustGet(addr)
-			if pfx.Overlaps(pc.prefix) && pfx.Bits() <= pc.prefix.Bits() {
-				// pfx covers path compressed prefix
-				allCoveredPathCompSet[addr] = true
-			}
-		}
-
-		addr++
-	}
-
-	// ###############################################################
-	// 3. collect all children in n covered by prefix
-	// ###############################################################
-
-	allCoveredChildSet := make([]bool, maxNodeChildren)
-
-	addr = 0
-
-	for {
-		if addr, ok = n.children.NextSet(addr); !ok {
-			break
-		}
-
-		// host addrs are sorted in indexRank order
-		if addr > pfxLastAddr {
-			break
-		}
-
-		if addr >= pfxFirstAddr {
-			allCoveredChildSet[addr] = true
-		}
-
-		addr++
-	}
-
-	// #########################################################################
-	// 4. yield covered indices, pathcomp prefixes and childs in CIDR sort order
-	// #########################################################################
-
-	// yield indices, pathcomp prefixes and childs in CIDR sort order
-	var lower, upper uint
-
-	for _, idx := range allCoveredIndices {
-		pfxAddr, _ := idxToPfx(idx)
-		upper = uint(pfxAddr)
-
-		// for all pathcomp and child items < pfxAddr
-		for addr := lower; addr < upper; addr++ {
-			// either pathcomp or children match this addr, but not possible for both
-
-			if allCoveredPathCompSet[addr] {
-				pc := n.pathcomp.MustGet(addr)
-				if !yield(pc.prefix, pc.value) {
-					return false
-				}
-			}
-
-			if allCoveredChildSet[addr] {
-				// yield this child rec-descent, if matched
-				c := n.children.MustGet(addr)
-				path[depth] = byte(addr)
-				if !c.allRecSorted(path, depth+1, is4, yield) {
-					return false
-				}
-			}
-		}
-
-		// yield the prefix for this idx
-		cidr, _ := cidrFromPath(path, depth, is4, idx)
-		if !yield(cidr, n.prefixes.MustGet(idx)) {
-			return false
-		}
-
-		// forward lower bound for next round
-		lower = upper
-	}
-
-	// #####################################################
-	// 5. yield the rest of pathcomp and child items, if any
-	// #####################################################
-
-	// yield the rest of pathcomp and child items, if any
-	for addr := lower; addr < maxNodeChildren; addr++ {
-		if allCoveredPathCompSet[addr] {
-			pc := n.pathcomp.MustGet(addr)
-			if !yield(pc.prefix, pc.value) {
-				return false
-			}
-		}
-
-		if allCoveredChildSet[addr] {
-			c := n.children.MustGet(addr)
-			path[depth] = byte(addr)
-			if !c.allRecSorted(path, depth+1, is4, yield) {
-				return false
-			}
-		}
-	}
-
-	return true
-}
-
-// unionRec combines two nodes, changing the receiver node.
-// If there are duplicate entries, the value is taken from the other node.
-// Count duplicate entries to adjust the t.size struct members.
-func (n *node[V]) unionRec(o *node[V], depth int) (duplicates int) {
-	allIndices := o.prefixes.AsSlice(make([]uint, 0, maxNodePrefixes))
-
-	// for all prefixes in other node do ...
-	for i, oIdx := range allIndices {
-		// insert/overwrite prefix/value from oNode to nNode
-		if exists := n.prefixes.InsertAt(oIdx, o.prefixes.Items[i]); exists {
-			// this prefix is duplicate in n and o
-			duplicates++
-		}
-	}
-
-	allPathCompAddrs := o.pathcomp.AsSlice(make([]uint, 0, maxNodeChildren))
-	// for all pathcomp items in other node do ...
-	for i, addr := range allPathCompAddrs {
-		oPCItem := o.pathcomp.Items[i]
-
-		// get n child with same addr, if exists insert prefix at depth
-		if nc, ok := n.children.Get(addr); ok {
-			if nc.insertAtDepth(oPCItem.prefix, oPCItem.value, depth+1) {
-				// this prefix is duplicate in n and o
-				duplicates++
-			}
-			continue
-		}
-
-		// no child found, look for path compressed item in slot
-		if nPCItem, ok := n.pathcomp.Get(addr); ok {
-			if nPCItem.prefix == oPCItem.prefix {
-				nPCItem.value = oPCItem.value
-				// this prefix is duplicate in n and o
-				duplicates++
-				continue
-			}
-
-			// free this pathcomp slot ...
-			// insert new intermdiate child ...
-			// shuffle down existing path-compressed prefix
-			// union other path-compressed prefix
-			n.pathcomp.DeleteAt(addr)
-
-			nc := new(node[V])
-			n.children.InsertAt(addr, nc)
-
-			// shuffle down
-			_ = nc.insertAtDepth(nPCItem.prefix, nPCItem.value, depth+1)
-
-			// union other
-			exists := nc.insertAtDepth(oPCItem.prefix, oPCItem.value, depth+1)
-			if exists {
-				duplicates++
-			}
-
-			continue
-		}
-
-		// no child nor pathcomp, insert as path compressed
-		n.pathcomp.InsertAt(addr, oPCItem)
-	}
-
-	allChildAddrs := o.children.AsSlice(make([]uint, 0, maxNodeChildren))
-
-	// for all children in other node do ...
-	for i, addr := range allChildAddrs {
-		oc := o.children.Items[i]
-
-		// get n pathcomp with same addr
-		if nPCItem, ok := n.pathcomp.Get(addr); ok {
-			// free this pathcomp slot ...
-			// insert new intermdiate child ...
-			// shuffle down existing path-compressed prefix
-			// union other child
-			n.pathcomp.DeleteAt(addr)
-
-			nc := new(node[V])
-			n.children.InsertAt(addr, nc)
-
-			// shuffle down
-			_ = nc.insertAtDepth(nPCItem.prefix, nPCItem.value, depth+1)
-
-			duplicates += nc.unionRec(oc, depth+1)
-			continue
-		}
-
-		// get n child with same addr,
-		if nc, ok := n.children.Get(addr); !ok {
-			// insert cloned child from oNode into nNode
-			n.children.InsertAt(addr, oc.cloneRec())
-		} else {
-			// both nodes have child with addr, call union rec-descent
-			duplicates += nc.unionRec(oc, depth+1)
-		}
-	}
-
-	return duplicates
-}
-
 // cloneRec, clones the node recursive.
 func (n *node[V]) cloneRec() *node[V] {
 	c := new(node[V])
@@ -484,38 +239,27 @@ func (n *node[V]) cloneRec() *node[V] {
 		return c
 	}
 
-	c.prefixes.BitSet = n.prefixes.BitSet.Clone()     // deep
-	c.prefixes.Items = slices.Clone(n.prefixes.Items) // values, shallow copy
+	// shallow
+	c.prefixes = *(n.prefixes.Clone())
 
 	// deep copy if V implements Cloner[V]
 	for i, v := range c.prefixes.Items {
-		if v, ok := any(v).(Cloner[V]); ok {
-			c.prefixes.Items[i] = v.Clone()
-		} else {
-			break
-		}
+		c.prefixes.Items[i] = cloneValue(v)
 	}
 
-	c.children.BitSet = n.children.BitSet.Clone()     // deep
-	c.children.Items = slices.Clone(n.children.Items) // children, shallow copy
+	// shallow
+	c.children = *(n.children.Clone())
 
-	// deep copy of children
-	for i, child := range c.children.Items {
-		c.children.Items[i] = child.cloneRec()
-	}
-
-	c.pathcomp.BitSet = n.pathcomp.BitSet.Clone()     // deep
-	c.pathcomp.Items = slices.Clone(n.pathcomp.Items) // values, shallow copy
-
-	// deep copy
-	for i, pc := range c.pathcomp.Items {
-		item := *pc
-
-		// deep copy if V implements Cloner[V]
-		if v, ok := any(item.value).(Cloner[V]); ok {
-			item.value = v.Clone()
+	// deep copy of nodes and leaves
+	for i, k := range c.children.Items {
+		switch k := k.(type) {
+		case *node[V]:
+			// clone the child node rec-descent
+			c.children.Items[i] = k.cloneRec()
+		case *leaf[V]:
+			// deep copy if V implements Cloner[V]
+			c.children.Items[i] = k.cloneLeaf()
 		}
-		c.pathcomp.Items[i] = &item
 	}
 
 	return c
@@ -533,37 +277,35 @@ func (n *node[V]) allRec(
 	is4 bool,
 	yield func(netip.Prefix, V) bool,
 ) bool {
-	allIndices := n.prefixes.AsSlice(make([]uint, 0, maxNodePrefixes))
-
 	// for all prefixes in this node do ...
+	allIndices := n.prefixes.AsSlice(make([]uint, 0, maxNodePrefixes))
 	for _, idx := range allIndices {
 		cidr, _ := cidrFromPath(path, depth, is4, idx)
 
-		// make the callback for this prefix
+		// callback for this prefix and val
 		if !yield(cidr, n.prefixes.MustGet(idx)) {
 			// early exit
 			return false
 		}
 	}
 
-	// for all path compressed items do ...
-	for _, pc := range n.pathcomp.Items {
-		// make the callback for this prefix
-		if !yield(pc.prefix, pc.value) {
-			// early exit
-			return false
-		}
-	}
-
+	// for all children (nodes and leaves) in this node do ...
 	allChildAddrs := n.children.AsSlice(make([]uint, 0, maxNodeChildren))
-	// for all children in this node do ...
 	for i, addr := range allChildAddrs {
-		child := n.children.Items[i]
-		path[depth] = byte(addr)
-
-		if !child.allRec(path, depth+1, is4, yield) {
-			// early exit
-			return false
+		switch k := n.children.Items[i].(type) {
+		case *node[V]:
+			// rec-descent with this node
+			path[depth] = byte(addr)
+			if !k.allRec(path, depth+1, is4, yield) {
+				// early exit
+				return false
+			}
+		case *leaf[V]:
+			// callback for this leaf
+			if !yield(k.prefix, k.value) {
+				// early exit
+				return false
+			}
 		}
 	}
 
@@ -582,70 +324,294 @@ func (n *node[V]) allRecSorted(
 	is4 bool,
 	yield func(netip.Prefix, V) bool,
 ) bool {
+	// get slice of all child octets, sorted by addr
+	allChildAddrs := n.children.AsSlice(make([]uint, 0, maxNodeChildren))
+
 	// get slice of all indexes, sorted by idx
 	allIndices := n.prefixes.AsSlice(make([]uint, 0, maxNodePrefixes))
 
 	// sort indices in CIDR sort order
 	slices.SortFunc(allIndices, cmpIndexRank)
 
-	// get all the bits in fast addressable form as a set of bool.
-	allChildSet := n.children.AsSet(make([]bool, 0, maxNodeChildren))
-	allPathCompSet := n.pathcomp.AsSet(make([]bool, 0, maxNodeChildren))
+	childCursor := 0
 
-	// yield indices, pathcomp prefixes and childs in CIDR sort order
-	var lower, upper uint
+	// yield indices and childs in CIDR sort order
+	for _, pfxIdx := range allIndices {
+		pfxOctet, _ := idxToPfx(pfxIdx)
 
-	for _, idx := range allIndices {
-		pfxAddr, _ := idxToPfx(idx)
-		upper = uint(pfxAddr)
+		// yield all childs before idx
+		for j := childCursor; j < len(allChildAddrs); j++ {
+			childAddr := allChildAddrs[j]
 
-		// for all pathcomp and child items < pfxAddr
-		for addr := lower; addr < upper; addr++ {
-			// either pathcomp or children match this addr, but not possible for both
+			if childAddr >= uint(pfxOctet) {
+				break
+			}
 
-			if allPathCompSet[addr] {
-				pc := n.pathcomp.MustGet(addr)
-				if !yield(pc.prefix, pc.value) {
+			// yield the node (rec-descent) or leaf
+			switch k := n.children.Items[j].(type) {
+			case *node[V]:
+				path[depth] = byte(childAddr)
+				if !k.allRecSorted(path, depth+1, is4, yield) {
+					return false
+				}
+			case *leaf[V]:
+				if !yield(k.prefix, k.value) {
 					return false
 				}
 			}
 
-			if allChildSet[addr] {
-				// yield this child rec-descent, if matched
-				c := n.children.MustGet(addr)
-				path[depth] = byte(addr)
-				if !c.allRecSorted(path, depth+1, is4, yield) {
-					return false
-				}
-			}
-
+			childCursor++
 		}
 
 		// yield the prefix for this idx
-		cidr, _ := cidrFromPath(path, depth, is4, idx)
-		if !yield(cidr, n.prefixes.MustGet(idx)) {
+		cidr, _ := cidrFromPath(path, depth, is4, pfxIdx)
+		// n.prefixes.Items[i] not possible after sorting allIndices
+		if !yield(cidr, n.prefixes.MustGet(pfxIdx)) {
 			return false
 		}
-
-		// forward lower bound for next round
-		lower = upper
 	}
 
-	// yield the rest of pathcomp and child items, if any
-	for addr := lower; addr < maxNodeChildren; addr++ {
-		// either pathcomp or children match this addr, but not possible for both
-
-		if allPathCompSet[addr] {
-			pc := n.pathcomp.MustGet(addr)
-			if !yield(pc.prefix, pc.value) {
+	// yield the rest of leaves and nodes (rec-descent)
+	for j := childCursor; j < len(allChildAddrs); j++ {
+		addr := allChildAddrs[j]
+		switch k := n.children.Items[j].(type) {
+		case *node[V]:
+			path[depth] = byte(addr)
+			if !k.allRecSorted(path, depth+1, is4, yield) {
+				return false
+			}
+		case *leaf[V]:
+			if !yield(k.prefix, k.value) {
 				return false
 			}
 		}
+	}
 
-		if allChildSet[addr] {
-			c := n.children.MustGet(addr)
+	return true
+}
+
+// unionRec combines two nodes, changing the receiver node.
+// If there are duplicate entries, the value is taken from the other node.
+// Count duplicate entries to adjust the t.size struct members.
+func (n *node[V]) unionRec(o *node[V], depth int) (duplicates int) {
+	// for all prefixes in other node do ...
+	allIndices := o.prefixes.AsSlice(make([]uint, 0, maxNodePrefixes))
+	for i, oIdx := range allIndices {
+		// insert/overwrite prefix/value from oNode to nNode
+		exists := n.prefixes.InsertAt(oIdx, o.prefixes.Items[i])
+
+		// this prefix is duplicate in n and o
+		if exists {
+			duplicates++
+		}
+	}
+
+	// for all child addrs in other node do ...
+	allOtherChildAddrs := o.children.AsSlice(make([]uint, 0, maxNodeChildren))
+LOOP:
+	for i, addr := range allOtherChildAddrs {
+		//  6 possible combinations for this child and other child child
+		//
+		//  THIS, OTHER:
+		//  ----------
+		//  NULL, node  <-- easy,    insert at cloned node
+		//  NULL, leaf  <-- easy,    insert at cloned leaf
+		//  node, node  <-- easy,    union rec-descent
+		//  node, leaf  <-- easy,    insert other cloned leaf at depth+1
+		//  leaf, node  <-- complex, push this leaf down, union rec-descent
+		//  leaf, leaf  <-- complex, push this leaf down, insert other cloned leaf at depth+1
+		//
+		// try to get child at same addr from n
+		thisChild, thisExists := n.children.Get(addr)
+		if !thisExists {
+			switch otherChild := o.children.Items[i].(type) {
+
+			case *node[V]: // NULL, node
+				if !thisExists {
+					n.children.InsertAt(addr, otherChild.cloneRec())
+					continue LOOP
+				}
+
+			case *leaf[V]: // NULL, leaf
+				if !thisExists {
+					n.children.InsertAt(addr, otherChild.cloneLeaf())
+					continue LOOP
+				}
+			}
+		}
+
+		switch otherChild := o.children.Items[i].(type) {
+
+		case *node[V]:
+			switch this := thisChild.(type) {
+
+			case *node[V]: // node, node
+				// both childs have node in octet, call union rec-descent on child nodes
+				duplicates += this.unionRec(otherChild, depth+1)
+				continue LOOP
+
+			case *leaf[V]: // leaf, node
+				// create new node
+				nc := new(node[V])
+
+				// push this leaf down
+				nc.insertAtDepth(this.prefix, this.value, depth+1)
+
+				// insert new node at current addr
+				n.children.InsertAt(addr, nc)
+
+				// union rec-descent new node with other node
+				duplicates += nc.unionRec(otherChild, depth+1)
+				continue LOOP
+			}
+
+		case *leaf[V]:
+			switch this := thisChild.(type) {
+
+			case *node[V]: // node, leaf
+				clonedLeaf := otherChild.cloneLeaf()
+				if this.insertAtDepth(clonedLeaf.prefix, clonedLeaf.value, depth+1) {
+					duplicates++
+				}
+				continue LOOP
+
+			case *leaf[V]: // leaf, leaf
+				// create new node
+				nc := new(node[V])
+
+				// push this leaf down
+				nc.insertAtDepth(this.prefix, this.value, depth+1)
+
+				// insert at depth cloned leaf
+				clonedLeaf := otherChild.cloneLeaf()
+				if nc.insertAtDepth(clonedLeaf.prefix, clonedLeaf.value, depth+1) {
+					duplicates++
+				}
+
+				// insert the new node at current addr
+				n.children.InsertAt(addr, nc)
+				continue LOOP
+			}
+		}
+	}
+
+	return duplicates
+}
+
+// eachLookupPrefix does an all prefix match in the 8-bit (stride) routing table
+// at this depth and calls yield() for any matching CIDR.
+func (n *node[V]) eachLookupPrefix(octets []byte, depth int, is4 bool, pfxLen int, yield func(netip.Prefix, V) bool) (ok bool) {
+	var path [16]byte
+	copy(path[:], octets)
+
+	if n.prefixes.Len() == 0 {
+		return true
+	}
+
+	// backtracking the CBT
+	for idx := pfxToIdx(octets[depth], pfxLen); idx > 0; idx >>= 1 {
+		if n.prefixes.Test(idx) {
+			val := n.prefixes.MustGet(idx)
+			cidr, _ := cidrFromPath(path, depth, is4, idx)
+
+			if !yield(cidr, val) {
+				return false
+			}
+		}
+	}
+
+	return true
+}
+
+// eachSubnet calls yield() for any covered CIDR by parent prefix in natural CIDR sort order.
+func (n *node[V]) eachSubnet(octets []byte, depth int, is4 bool, pfxLen int, yield func(netip.Prefix, V) bool) bool {
+	// octets as array, needed below more than once
+	var path [16]byte
+	copy(path[:], octets)
+
+	pfxFirstAddr := uint(octets[depth])
+	pfxLastAddr := uint(octets[depth] | ^netMask(pfxLen))
+
+	// 1. collect all indices in n covered by prefix
+
+	allCoveredIndices := make([]uint, 0, maxNodePrefixes)
+	for _, idx := range n.prefixes.AsSlice(make([]uint, 0, maxNodePrefixes)) {
+		thisOctet, thisPfxLen := idxToPfx(idx)
+
+		thisFirstAddr := uint(thisOctet)
+		thisLastAddr := uint(thisOctet | ^netMask(thisPfxLen))
+
+		if thisFirstAddr >= pfxFirstAddr && thisLastAddr <= pfxLastAddr {
+			allCoveredIndices = append(allCoveredIndices, idx)
+		}
+	}
+
+	// sort indices in CIDR sort order
+	slices.SortFunc(allCoveredIndices, cmpIndexRank)
+
+	// 2. collect all covered child addrs by prefix
+
+	allCoveredChildAddrs := make([]uint, 0, maxNodeChildren)
+	for _, addr := range n.children.AsSlice(make([]uint, 0, maxNodeChildren)) {
+		if addr >= pfxFirstAddr && addr <= pfxLastAddr {
+			allCoveredChildAddrs = append(allCoveredChildAddrs, addr)
+		}
+	}
+
+	// 3. yield covered indices, pathcomp prefixes and childs in CIDR sort order
+
+	addrCursor := 0
+
+	// yield indices and childs in CIDR sort order
+	for _, pfxIdx := range allCoveredIndices {
+		pfxOctet, _ := idxToPfx(pfxIdx)
+
+		// yield all childs before idx
+		for j := addrCursor; j < len(allCoveredChildAddrs); j++ {
+			addr := allCoveredChildAddrs[j]
+			if addr >= uint(pfxOctet) {
+				break
+			}
+
+			// yield the node or leaf?
+			switch k := n.children.MustGet(addr).(type) {
+
+			case *node[V]:
+				path[depth] = byte(addr)
+				if !k.allRecSorted(path, depth+1, is4, yield) {
+					return false
+				}
+
+			case *leaf[V]:
+				if !yield(k.prefix, k.value) {
+					return false
+				}
+			}
+
+			addrCursor++
+		}
+
+		// yield the prefix for this idx
+		cidr, _ := cidrFromPath(path, depth, is4, pfxIdx)
+		// n.prefixes.Items[i] not possible after sorting allIndices
+		if !yield(cidr, n.prefixes.MustGet(pfxIdx)) {
+			return false
+		}
+	}
+
+	// yield the rest of leaves and nodes (rec-descent)
+	for _, addr := range allCoveredChildAddrs[addrCursor:] {
+		// yield the node or leaf?
+		switch k := n.children.MustGet(addr).(type) {
+
+		case *node[V]:
 			path[depth] = byte(addr)
-			if !c.allRecSorted(path, depth+1, is4, yield) {
+			if !k.allRecSorted(path, depth+1, is4, yield) {
+				return false
+			}
+
+		case *leaf[V]:
+			if !yield(k.prefix, k.value) {
 				return false
 			}
 		}
