+++ worstcase_test.go
@@ -9,17 +9,17 @@ var (
 	worstCaseProbeIP4 = mpa("255.255.255.255")
 
 	worstCasePfxsIP4 = []netip.Prefix{
-		mpp("0.0.0.0/0"), // matching prefix
+		mpp("0.0.0.0/1"),
 		mpp("255.0.0.0/9"),
 		mpp("255.255.0.0/17"),
 		mpp("255.255.255.0/25"),
-		mpp("255.255.255.254/32"),
+		mpp("255.255.255.255/32"), // matching prefix
 	}
 
 	worstCaseProbeIP6 = mpa("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff")
 
 	worstCasePfxsIP6 = []netip.Prefix{
-		mpp("::/0"), // matching prefix
+		mpp("::/1"),
 		mpp("ff00::/9"),
 		mpp("ffff::/17"),
 		mpp("ffff:ff00::/25"),
@@ -36,6 +36,7 @@ var (
 		mpp("ffff:ffff:ffff:ffff:ffff:ffff:ffff::/113"),
 		mpp("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ff00/121"),
 		mpp("ffff:ffff:ffff:ffff:ffff:ffff:ffff:fffe/128"),
+		mpp("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff/128"),
 	}
 )
 
@@ -65,7 +66,7 @@ func TestWorstCase(t *testing.T) {
 			tbl.Insert(p, p.String())
 		}
 
-		tbl.Delete(mpp("0.0.0.0/0")) // delete matching prefix
+		tbl.Delete(mpp("255.255.255.255/32")) // delete matching prefix
 
 		want := false
 		ok := tbl.Contains(worstCaseProbeIP4)
@@ -88,6 +89,7 @@ func TestWorstCase(t *testing.T) {
 			t.Errorf("Contains, worst case match IP6, expected OK: %v, got: %v", want, ok)
 		}
 	})
+
 	t.Run("WorstCaseMissIP6", func(t *testing.T) {
 		t.Parallel()
 
@@ -96,74 +98,7 @@ func TestWorstCase(t *testing.T) {
 			tbl.Insert(p, p.String())
 		}
 
-		tbl.Delete(mpp("::/0")) // delete matching prefix
-
-		want := false
-		ok := tbl.Contains(worstCaseProbeIP6)
-		if ok != want {
-			t.Errorf("Contains, worst case miss IP6, expected OK: %v, got: %v", want, ok)
-		}
-	})
-}
-
-func TestWorstCasePC(t *testing.T) {
-	t.Parallel()
-
-	t.Run("WorstCaseMatchIP4", func(t *testing.T) {
-		t.Parallel()
-
-		tbl := new(Table2[string])
-		for _, p := range worstCasePfxsIP4 {
-			tbl.Insert(p, p.String())
-		}
-
-		want := true
-		ok := tbl.Contains(worstCaseProbeIP4)
-		if ok != want {
-			t.Errorf("Contains, worst case match IP4, expected OK: %v, got: %v", want, ok)
-		}
-	})
-
-	t.Run("WorstCaseMissIP4", func(t *testing.T) {
-		t.Parallel()
-
-		tbl := new(Table2[string])
-		for _, p := range worstCasePfxsIP4 {
-			tbl.Insert(p, p.String())
-		}
-
-		tbl.Delete(mpp("0.0.0.0/0")) // delete matching prefix
-
-		want := false
-		ok := tbl.Contains(worstCaseProbeIP4)
-		if ok != want {
-			t.Errorf("Contains, worst case miss IP4, expected OK: %v, got: %v", want, ok)
-		}
-	})
-
-	t.Run("WorstCaseMatchIP6", func(t *testing.T) {
-		t.Parallel()
-
-		tbl := new(Table2[string])
-		for _, p := range worstCasePfxsIP6 {
-			tbl.Insert(p, p.String())
-		}
-
-		want := true
-		ok := tbl.Contains(worstCaseProbeIP6)
-		if ok != want {
-			t.Errorf("Contains, worst case match IP6, expected OK: %v, got: %v", want, ok)
-		}
-	})
-	t.Run("WorstCaseMissIP6", func(t *testing.T) {
-		t.Parallel()
-
-		tbl := new(Table2[string])
-		for _, p := range worstCasePfxsIP6 {
-			tbl.Insert(p, p.String())
-		}
-
-		tbl.Delete(mpp("::/0")) // delete matching prefix
+		tbl.Delete(mpp("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff/128")) // delete matching prefix
 
 		want := false
 		ok := tbl.Contains(worstCaseProbeIP6)
@@ -174,7 +109,7 @@ func TestWorstCasePC(t *testing.T) {
 }
 
 func BenchmarkWorstCase(b *testing.B) {
-	b.Run("WorstCase IP4", func(b *testing.B) {
+	b.Run("WorstCaseMatchIP4", func(b *testing.B) {
 		tbl := new(Table[string])
 		for _, p := range worstCasePfxsIP4 {
 			tbl.Insert(p, p.String())
@@ -185,7 +120,7 @@ func BenchmarkWorstCase(b *testing.B) {
 		}
 	})
 
-	b.Run("WorstCase IP6", func(b *testing.B) {
+	b.Run("WorstCaseMatchIP6", func(b *testing.B) {
 		tbl := new(Table[string])
 		for _, p := range worstCasePfxsIP6 {
 			tbl.Insert(p, p.String())
@@ -195,26 +130,27 @@ func BenchmarkWorstCase(b *testing.B) {
 			_ = tbl.Contains(worstCaseProbeIP6)
 		}
 	})
-}
 
-func BenchmarkWorstCasePC(b *testing.B) {
-	b.Run("WorstCasePC IP4", func(b *testing.B) {
-		tbl := new(Table2[string])
-
+	b.Run("WorstCaseMissIP4", func(b *testing.B) {
+		tbl := new(Table[string])
 		for _, p := range worstCasePfxsIP4 {
 			tbl.Insert(p, p.String())
 		}
+
+		tbl.Delete(mpp("255.255.255.255/32")) // delete matching prefix
 
 		for range b.N {
 			_ = tbl.Contains(worstCaseProbeIP4)
 		}
 	})
 
-	b.Run("WorstCasePC IP6", func(b *testing.B) {
-		tbl := new(Table2[string])
+	b.Run("WorstCaseMissIP6", func(b *testing.B) {
+		tbl := new(Table[string])
 		for _, p := range worstCasePfxsIP6 {
 			tbl.Insert(p, p.String())
 		}
+
+		tbl.Delete(mpp("ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff/128")) // delete matching prefix
 
 		for range b.N {
 			_ = tbl.Contains(worstCaseProbeIP6)
